+++ lib_acl_cpp/include/acl_cpp/db/sqlite_manager.hpp
@@ -12,25 +12,25 @@ public:
 
 	/**
 	* @param dbfile {const char*} sqlite 数据库的数据文件
-	* @param dblimit {int} 数据库连接池最大连接数限制
+	* @param dblimit {size_t} 数据库连接池最大连接数限制
 	* @return {sqlite_manager&}
 	 */
-	sqlite_manager& add(const char* dbfile, int dblimit);
+	sqlite_manager& add(const char* dbfile, size_t dblimit);
 
 protected:
 	/**
 	 * 基类 connect_manager 虚函数的实现
 	 * @param addr {const char*} 服务器监听地址，格式：ip:port
-	 * @param count {int} 连接池的大小限制
+	 * @param count {size_t} 连接池的大小限制
 	 * @param idx {size_t} 该连接池对象在集合中的下标位置(从 0 开始)
 	 * @return {connect_pool*} 返回创建的连接池对象
 	 */
-	connect_pool* create_pool(const char* addr, int count, size_t idx);
+	connect_pool* create_pool(const char* addr, size_t count, size_t idx);
 
 private:
 	// sqlite 数据文件名
 	char* dbfile_;
-	int   dblimit_;
+	size_t dblimit_;
 };
 
 } // namespace acl
